+++ src/plone/restapi/types/configure.zcml
@@ -17,7 +17,7 @@
     <adapter factory=".adapters.ListJsonSchemaProvider" />
     <adapter
         zcml:condition="installed z3c.relationfield"
-        factory=".adapters.ChoiceslessRelationListSchemaProvider"
+        factory=".z3crelationadapter.ChoiceslessRelationListSchemaProvider"
         name="IRelatedItems.relatedItems" />
     <adapter factory=".adapters.SetJsonSchemaProvider" />
     <adapter factory=".adapters.TupleJsonSchemaProvider" />
